remove react-hooks testing library, use dummy component to test hook
