+++ client/package.json
@@ -45,7 +45,6 @@
   "devDependencies": {
     "@testing-library/jest-dom": "^5.4.5",
     "@testing-library/react": "^15.0.7",
-    "@testing-library/react-hooks": "8.0.1",
     "@types/jest": "^29.5.2",
     "@types/node": "^20",
     "@types/react": "^18",
